Add graph production and SPARTA run, as well as remap several paths.
This is a large change that folds in the work done for
add-blazegraph-runner. In additions to now producing many primary and
secondary products in a mega-step, we also move the locations of
several drop locations to be more consistent. (E.g. reports/ being
used and products/ getting sub-directories.) Unfortunate growth in
repetition of copy commands.

TODO: we still need TTLs to be copied over to products/.